+++ zipline-kotlin-plugin/src/main/kotlin/app/cash/zipline/kotlin/AddAdapterArgumentRewriter.kt
@@ -13,8 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+@file:OptIn(DeprecatedForRemovalCompilerApi::class)
+
 package app.cash.zipline.kotlin
 
+import org.jetbrains.kotlin.DeprecatedForRemovalCompilerApi
 import org.jetbrains.kotlin.backend.common.ScopeWithIr
 import org.jetbrains.kotlin.backend.common.extensions.IrPluginContext
 import org.jetbrains.kotlin.ir.declarations.IrDeclarationParent
@@ -79,7 +82,7 @@ internal class AddAdapterArgumentRewriter(
     ).adapterExpression(bridgedInterfaceType as IrSimpleType)
 
     return irCall(original, rewrittenFunction).apply {
-      putValueArgument(valueArgumentsCount - 1, adapterExpression)
+      arguments += adapterExpression
       patchDeclarationParents(declarationParent)
     }
   }