+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * In-development snapshots are now published to the Central Portal Snapshots repository at https://central.sonatype.com/repository/maven-snapshots/.
 
+ * Upgrade: [Kotlin 2.2.0][kotlin_2_2_0].
+
 
 ## [1.21.1] - 2025-07-17
 [1.21.1]: https://github.com/cashapp/zipline/releases/tag/1.21.1
@@ -617,6 +619,7 @@ Initial release.
 [kotlin_1_9_23]: https://github.com/JetBrains/kotlin/releases/tag/v1.9.23
 [kotlin_2_0_0]: https://github.com/JetBrains/kotlin/releases/tag/v2.0.0
 [kotlin_2_1_21]: https://github.com/JetBrains/kotlin/releases/tag/v2.1.21
+[kotlin_2_2_0]: https://github.com/JetBrains/kotlin/releases/tag/v2.2.0
 [kotlin_serialization_1_4_0]: https://github.com/Kotlin/kotlinx.serialization/releases/tag/v1.4.0
 [kotlin_serialization_1_5_0]: https://github.com/Kotlin/kotlinx.serialization/releases/tag/v1.5.0
 [kotlin_serialization_1_5_1]: https://github.com/Kotlin/kotlinx.serialization/releases/tag/v1.5.1